requirements.txt
dj-database-url==0.5.0
Django==3.0.4
django-ckeditor==5.9.0
django-cors-headers==3.3.0
django-crispy-forms==1.4.0
django-crum==0.7.0
django-debug-toolbar==3.2
django-durationfield==0.5.1
django-extensions==3.0.3
django-filter==2.3.0
django-grappelli==2.14.2
django-jet==1.0.8
django-js-asset==1.2.2
django-material==1.7.1
django-userforeignkey==0.3.0
django-utils-six==2.0
django-widget-tweaks==1.4.8
djangorestframework==3.11.0
fontawesome==5.10.1.post1
graphviz==0.13.2
gunicorn==20.0.4
html5lib==1.0.1
httpie==2.4.0
idna==2.10
importlib-metadata==3.4.0
isort==4.3.21
jdcal==1.4.1
Jinja2==2.11.3
lazy-object-proxy==1.4.3
Markdown==3.1.1
MarkupSafe==1.1.1
mccabe==0.6.1
MonthDelta==0.9.1
openpyxl==3.0.5
parse==1.19.0
parse-type==0.5.2
Pillow==8.1.2
psycopg2-binary==2.8.6
pycodestyle==2.6.0
pydotplus==2.0.2
Pygments==2.9.0
pylint==2.4.4
python-dateutil==2.7.2
python-decouple==3.3
pytz==2019.3
sorl-thumbnail==12.7.0
#sqlparse==0.3.1
#toml==0.10.2
#typed-ast==1.4.2
#typing-extensions==3.7.4.3
#urllib3==1.26.4
#webencodings==0.5.1
whitenoise==4.1
#wrapt==1.11.1
#xhtml2pdf==0.2.4
#xlrd==1.2.0
#zipp==3.4.0
